Bleeding jumbo

+++ src/unrarfilter.c
@@ -24,7 +24,6 @@
 
 #include "unrar.h"
 #include "unrarfilter.h"
-#include "memdbg.h"
 
 void rar_filter_array_init(rar_filter_array_t *filter_a)
 {
@@ -34,7 +33,7 @@ void rar_filter_array_init(rar_filter_array_t *filter_a)
 
 void rar_filter_array_reset(rar_filter_array_t *filter_a)
 {
-	int i;
+	size_t i;
 
 	if (!filter_a) {
 		return;
